requirements.txt
Django==1.11.17
django-cors-headers==2.4.0
djangorestframework==3.9.0
pytz==2018.7
